+++ src/formulas/mixing_ratio.rs
@@ -3,7 +3,7 @@
 //! Mixing ratio is the ratio of the mass of a variable atmospheric constituent to the mass
 //! of dry air ([AMETSOC Glossary](https://glossary.ametsoc.org/wiki/Mixing_ratio)).
 
-use crate::formula::Formula2;
+use crate::Formula2;
 use crate::quantities::{AtmosphericPressure, MixingRatio, ThermodynamicQuantity, VapourPressure};
 use crate::Float;
 use crate::{constants::EPSILON, errors::InputError};
